backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.12.0
python-dotenv
bcrypt==3.2.0
matplotlib==3.4.3
imutils==0.5.4
scipy==1.7.1
opencv-python==4.5.3.56
dlib==19.22.0
pandas==1.3.3
scikit-learn==0.24.2
